backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pydantic==2.5.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6